functional completion of task, service, and endpoints

+++ Week1/TaskManagement/Program.cs
@@ -28,7 +28,7 @@ app.UseHttpsRedirection();
 TaskService service = new TaskService();
 
 //Get all tasks with optional filtering - Query parameters: isCompleted, priority, dueBefore
-app.MapGet("/api/tasks", ([FromBody] string sort, ILogger<Program> logger) =>
+app.MapGet("/api/tasks", ([FromQuery] string? sort, ILogger<Program> logger) =>
 {
     if(sort != "priority" && sort != "Priority" && sort != "Completed" && sort != "completed" && sort != "dueDate" && sort != "duedate") {
         return Results.BadRequest(new { success = false, data = "Invalid sort method [Priority, Completed, dueDate]", message = "Operation failed" });
@@ -50,24 +50,31 @@ app.MapGet("/api/tasks/{id}", (int id) =>
 });
 
 //Create new task
-app.MapPost("/api/tasks", ([FromBody] int id, ILogger<Program> logger) => 
-{ 
-    logger.LogInformation("Sort is: " + id);
-    return Results.Ok(new { success = true, data = service.updateTask(id), message = "Operation completed successfully"});
+app.MapPost("/api/tasks", (Tasc task) => 
+{
+    
+    return Results.Ok(new { success = true, data = service.addToList(task), message = "Operation completed successfully"});
 });
 
 //Update existing task
-app.MapPut("/api/tasks/{id}", (int id) => 
-{ 
-    return Results.Ok(new { success = true, data = service.updateTask(id), message = "Operation completed successfully"});
-});
+app.MapPut("/api/tasks/{id}", (int id, [FromBody] Tasc task) => 
+{
+    if(service.findTask(id) == null) {
+        return Results.BadRequest(new { success = false, data = "ID not found", message = "Operation failed" });
+    }
+
+    var updated = service.updateTasc(id, title: task.title, description: task.description, isCompleted: task.isCompleted, priority: task.priority, dueDate: task.dueDate?.ToString());
+
+    return Results.Ok(new { success = true, data = updated, message = "Operation completed successfully"});
+}); 
 
 //Delete task
 app.MapDelete("/api/tasks/{id}", (int id) => 
 { 
+    if(service.findTask(id) == null) {
+        return Results.BadRequest(new { success = false, data = "ID not found", message = "Operation failed" });
+    }
     return Results.Ok(new { success = true, data = service.deleteTask(id), message = "Operation completed successfully"});
 });
-
-
 
 app.Run();
